requirements.txt
androguard==3.3.5
anytree==2.4.3
apache-libcloud==2.4.0
appdirs==1.4.3
appnope==0.1.0
args==0.1.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==18.2.0
backcall==0.1.0
bcrypt==3.1.6
beancount==2.2.1
beautifulsoup4==4.6.3
black==18.9b0
bottle==0.12.16
cachetools==3.1.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
click==6.7
clint==0.5.1
colorama==0.4.1
cryptography==2.6.1
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
docker-py==1.10.6
docker-pycreds==0.4.0
fdroidserver==1.1.1
Flask==1.0.2
future==0.17.1
gitdb2==2.0.5
GitPython==2.1.11
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
lxml==4.5.1
more-itertools==8.4.0
packaging==20.4
plaid-python==7.1.0
pluggy==0.13.1
ply==3.11
prompt-toolkit==2.0.9
protobuf==3.6.1
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.1
pycparser==2.19
pydot==1.4.1
Pygments==2.3.1
pymongo==3.7.2
PyNaCl==1.3.0
pyparsing==2.3.1
pytest==4.3.1
python-dateutil==2.8.0
python-magic==0.4.15
python-vagrant==0.5.15
PyYAML==3.13
qrcode==6.1
requests==2.7.0
requests-oauthlib==1.2.0
rsa==4.0
ruamel.yaml==0.15.89
six==1.12.0
smmap2==2.0.5
toml==0.10.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.55.0
Werkzeug==0.15.1
xkcdpass==1.17.2
